+++ pwv_kpno/exceptions.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+# -*- coding: UTF-8 -*-
+
+#    This file is part of the pwv_kpno software package.
+#
+#    The pwv_kpno package is free software: you can redistribute it and/or
+#    modify it under the terms of the GNU General Public License as published
+#    by the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    The pwv_kpno package is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+#    Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with pwv_kpno. If not, see <http://www.gnu.org/licenses/>.
+
+
+"""Custom exceptions for ``pwv_kpno``."""
+
+
+class ModelingConfigError(Exception):
+    pass
